Improve Talus with quadrilateral fit
As the initial guesses of Y0 and Z0 are not reliable for the  talus a new function was developped to identify the superior face of the talus and its direction.
The criteria to identify the talus-calcaneus AS were modified to be based on the minimal curvature.
Also the fitCSATalus was corrected to ensure two positive gaussian curves are added.
The solution works fine for most of the test cases.