Correction and adaptation of TriPatellaFun
There were some minor bugs in the function.
In order to make it work on mesh composed larger triangles the parameters of the morphology operation have been modified. also the TriErodeMesh.m and the TriDilateMesh.m functions were modified.

Tri plan intersect now gives the area of each curves it identifies, and the calculation of the total area of the cross section accounting for potentially nested holes are is more general than before.