Improve FootSole [Work in Progress]
## Done

A few things were changed, now an initial superior to inferior axis (blue arrow) is computed from the centroid of the foot triangulation (black sphere) and the centroid of the convex hull of the foot (red sphere) which is located inferiorly relative to the one of the foot. The vector connecting the two spheres is made perpendicular to the principal inertia axis of the foot (posterior to anterior axis) to get an approximate vector (blue arrow) of the normal of the largest triangle.
![Foot](https://user-images.githubusercontent.com/29957158/74863851-ddaedb00-534e-11ea-8126-ef2f598dbe5e.JPG)

Then the convexhull is projected on plane (a first approximation of the sole plane) perpendicular to this vector (blue arrow). The largest triangle (area-wise) is identified on the projected version of the convex hull. This allows to eliminate the selection of a large triangle that can be found on the medial side of the foot, because the area of this triangle is drastically reduced when its projected on the first guess of the sole plane. However this could be used to the orientation of the medial to lateral vector.

With this method a correct blue plane is identified for all geometries of the test set.

### ToDo

Given that the points on the metatarsus are closer than with the calcaneus tip, the anterior to posterior orientation can be identified for the first inertia axis.

The next step I want to implement to get the calcaneus tip and the medial and lateral metatarsal bone tips (the three points of interest for the foot sole):

1.  Select from the longest edges those which are below a plan parallel to the one defined by the largest triangle (the start and end of those edges must be below that plan)
2.  Cluster the points defining the largest edges in 3 clusters
3.  Associate the cluster to the medial and lateral distal points of the triangles
4.  Get the points as the furthest one from the ones from the triangles -> this gives the points corresponding to the distal medial and lateral metatarsal bone tips
5.  Keep the proximal vertices of the triangle as the calcaneus tip