Working prototypes of the femur_get_correct_first_cs
Function to test putting back together a correct orientation of the femur.
General Idea
The idea is to exploit the fact that the distal epiphysis of the bone is relatively symmetrical relative to the diaphysis axis, while the one from
the proximal epiphysis is not because of the femoral head. So if you deform the femur along the second principal inertial deirection which is relatively medial to lateral, the centroid of the distal epiphysis should be closer to the first principal inertia axis than the one from the proximal epiphysis.